fftech-webaudit-saas/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.12
passlib[bcrypt]>=1.7
PyJWT>=2.8
APScheduler>=3.10
reportlab>=4.1
python-multipart>=0.0.9
requests>=2.32
